Added kinect_calibration_control subscriber to control saving images and exiting the calibration node.
In order to use the calibration node from other ROS nodes for automated calibration using a fixed kinect and calibration patterns held by robotic arms, an interface using ROS topic is added. With the keyboard input, it is only possible to do manual calibration. In specific, this topic is useful for the automated calibration of the kinect mounted to the upgraded versions of the PR2.