Skipping image if no ROI could be detected.
If there is a problem finding/calculating the undistorted points, the computeROI function will now return a failure. This will cause the calibrate function to skip the image for which no ROI was found. A WARN message is printed to the console.

This prevent the calibration routine from throwing a CV exception and crashing.