Added error handling for unusable images in file restore function.
If the number of detected points does not match the points on the pattern or if the detected color points do not match the detected IR points, the calibrate function throws a CV exception.

DepthCalibration and CameraCalibration now print an ERROR message and return false. The return value is handled and prevents the calibrate function to be run. The ERROR message suggests to remove the unusable images by hand and restart the calibration procedure.